fix: update test content for problem_si
Test include a bug in answer, which does not correlate
with the rest of the tests kit.

Input:
```
20
0 0 1 2 2 2 2 2 3 3 4 5 5 2 1 2 3 4 5 6
13
```
It terms of task, first 13 elements of the array
are sorted and participant's function should
return index of insertion of the first element
in unsorted part of the array. In that case it
would be arr[13] = 2 and transition will be look like
```
arr[13] = 2  -> arr[8] = 3
```
New key should be the same as problem statement says.

And after array state is
```
0 0 1 2 2 2 2 2 | 3 3 3 4 5 5 | 1 2 3 4 5 6
8
```

But before is was like
```
0 0 1 2 2 2 2 2 3 3 4 5 5 7 7 7 7 7 7 7
13
```

More over it is not clear where from these
7 comes from, because value 7 is not
represented in the original array.