fix: mismatch between number of expected args for testing program
In a previous version of problem_agf_begin.c, the program was expected
to take a pair of arguments (x and y) from test files for problem_agf,
but all test cases contained only one (x).

As I figured out, all tests cases rely on the assumption that
if (x < 0) => y = 0, then x >= 0, value of y doesn't matter.

Now y is set to a fixed value, according to the assumption above.